add Extras module
