+++ lemmy-2/modules/memes.py
@@ -28,14 +28,3 @@ class Memes(Module):
 	async def cmd_genjimain(self, message, args, kwargs):
 		genji_mains = self.load_data('genjimain')['links']
 		await self.client.send_message(message.channel, random.choice(genji_mains))
-
-	docs_8ball = {
-		'description': 'Gives a certified correct answer to your question'
-	}
-	async def cmd_8ball(self, message, args, kwargs):
-		responses = [ "It is certain.", "It is decidedly so.", "Without a doubt.", "Yes, definitely.",
-						"You may rely on it.", "As I see it, yes.", "Most likely.", "Outlook good.", "Yes.",
-						"Signs point to yes.", "Reply hazy try again.", "Ask again later.", "Better not tell you now.",
-						"Cannot predict now.", "Concentrate and ask again.", "Don't count on it.", "My reply is no.",
-						"My sources say no.", "Outlook not so good.", "Very doubtful." ]
-		await self.client.send_message(message.channel, message.author.mention + " :8ball: " + random.choice(responses))
